git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/alfresco/HEAD/root@44675 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261

+++ source/java/org/alfresco/repo/webdav/MoveMethod.java
@@ -205,7 +205,18 @@ public class MoveMethod extends HierarchicalMethod
             setHidden(destFileInfo.getNodeRef(), false);
             if (isMove)
             {
-                fileFolderService.delete(sourceNodeRef);
+                if (getDAVHelper().isRenameShuffle(destPath) && !getDAVHelper().isRenameShuffle(sourcePath))
+                {
+                    // if temporary or backup file already exists
+                    // don't delete source that is node with version history
+                    setHidden(sourceNodeRef, true);
+                    // As per the WebDAV spec, we make sure the node is unlocked once moved
+                    getDAVHelper().getLockService().unlock(sourceNodeRef);
+                }
+                else
+                {
+                    fileFolderService.delete(sourceNodeRef);
+                }
             }
         }
         // If this is a copy then the source is just copied to destination.
